refactor(ctf): resolve SonarCloud maintainability findings (#1689)
Clear the SonarCloud quality gate on the CTF workspace SPA — maintainability
code smells only, no behavior change (CTF contract + page suites unchanged).

Backend:
- Split the 2,058-line ctf/api/organizer_views.py into a ctf/api/organizer/
  package (_base + events/challenges/attachments/play/participants/ranges/
  notifications/scoreboard, each < 500 lines) and the 952-line
  ctf/api/serializers.py into a ctf/api/serializers/ package; __init__ re-exports
  keep every import path stable.
- Cut the "too many returns" (S1142) across the DRF views at the root: resolve/
  ownership/validation helpers raise a shared _CtfApiError that a single
  per-view except renders to the exact status/message (byte-identical responses,
  Retry-After preserved).
- Replace Any type hints in projections.py with the real model types; tighten the
  delete-helper callable type.

Frontend:
- Extract a duplicated query-key literal to a constant; split the 668-line
  ChallengeFormPage into ChallengeFormSections + challenge-form-state; reduce
  cognitive complexity in the Challenge/Event forms via extracted helpers;
  hoist nested ternaries into named values; invert negated conditions; guard
  format.ts stringification.

OpenAPI contract unchanged (api_contract --check clean, zero drift).
