+++ shifter/shifter_platform/ctf/api/serializers/__init__.py
@@ -1,0 +1,186 @@
+"""Serializers for the canonical CTF DRF API.
+
+Split by audience across ``participant`` (the ``/api/v1/ctf/me/*`` workspace read
+projections plus the participant play surface), ``organizer`` (event + challenge
+management), and ``organizer_ops`` (participant / range / notification /
+scoreboard management). This package re-exports every serializer so existing
+``from ctf.api.serializers import X`` imports keep resolving unchanged.
+"""
+
+from __future__ import annotations
+
+from ctf.api.serializers.organizer import (
+    ChallengeFileListResponseSerializer,
+    ChallengeFileMetaSerializer,
+    ChallengeFileUploadResultSerializer,
+    ChallengeFileUploadSerializer,
+    ChallengeHintSerializer,
+    ChallengeListResponseSerializer,
+    ChallengeMutationResultSerializer,
+    ChallengeSummarySerializer,
+    ChallengeWriteSerializer,
+    CtfScenarioListResponseSerializer,
+    CtfScenarioRefSerializer,
+    DeleteSuccessSerializer,
+    EventDetailSerializer,
+    EventListResponseSerializer,
+    EventMutationResultSerializer,
+    EventSummarySerializer,
+    EventWriteSerializer,
+    FileDownloadResponseSerializer,
+    FlagCreateResultSerializer,
+    FlagWriteSerializer,
+    ForceDeleteEventRequestSerializer,
+    ForceDeleteEventResultSerializer,
+    HintListResponseSerializer,
+    HintWriteSerializer,
+    OrganizerChallengeDetailSerializer,
+    PrerequisiteCreateResultSerializer,
+    PrerequisiteListResponseSerializer,
+    PrerequisiteSerializer,
+    PrerequisiteWriteSerializer,
+)
+from ctf.api.serializers.organizer_ops import (
+    AssignBracketRequestSerializer,
+    AssignBracketResultSerializer,
+    EmailTemplateResponseSerializer,
+    EmailTemplateRevertResultSerializer,
+    EmailTemplateWriteSerializer,
+    NotificationAnnounceRequestSerializer,
+    NotificationAnnounceResultSerializer,
+    NotificationListItemSerializer,
+    NotificationListResponseSerializer,
+    NotificationSendResultSerializer,
+    OrganizerScoreboardResponseSerializer,
+    ParticipantDeleteResultSerializer,
+    ParticipantDetailSerializer,
+    ParticipantImportedItemSerializer,
+    ParticipantImportErrorSerializer,
+    ParticipantImportResultSerializer,
+    ParticipantImportSerializer,
+    ParticipantInviteResultSerializer,
+    ParticipantInviteSerializer,
+    ParticipantListResponseSerializer,
+    ParticipantRangeActionResultSerializer,
+    ParticipantSummarySerializer,
+    PublicScoreboardResponseSerializer,
+    RangeAccessResponseSerializer,
+    RangeListItemSerializer,
+    RangeListResponseSerializer,
+    RangeProvisionQueuedSerializer,
+    RangeRecoveryRequestSerializer,
+    RangeRecoveryResultSerializer,
+    RangeStatusResponseSerializer,
+    ResendInviteResultSerializer,
+    ScoreTimelineResponseSerializer,
+    SendInvitationsResultSerializer,
+    SparePoolRequestSerializer,
+    SpareProvisionResultSerializer,
+)
+from ctf.api.serializers.participant import (
+    ChallengeConnectionInfoSerializer,
+    ChallengeRatingSerializer,
+    ParticipantChallengeDetailSerializer,
+    ParticipantChallengeFileSerializer,
+    ParticipantChallengeListItemSerializer,
+    ParticipantCurrentEventSerializer,
+    ParticipantEventSerializer,
+    ParticipantHintSerializer,
+    ParticipantSelfSerializer,
+    ParticipantTeamMemberSerializer,
+    ParticipantTeamSerializer,
+    RateChallengeRequestSerializer,
+    RateChallengeResultSerializer,
+    SubmissionListItemSerializer,
+    SubmissionListResponseSerializer,
+    SubmitFlagRequestSerializer,
+    SubmitFlagResultSerializer,
+    UseHintRequestSerializer,
+    UseHintResultSerializer,
+)
+
+__all__ = [
+    "AssignBracketRequestSerializer",
+    "AssignBracketResultSerializer",
+    "ChallengeConnectionInfoSerializer",
+    "ChallengeFileListResponseSerializer",
+    "ChallengeFileMetaSerializer",
+    "ChallengeFileUploadResultSerializer",
+    "ChallengeFileUploadSerializer",
+    "ChallengeHintSerializer",
+    "ChallengeListResponseSerializer",
+    "ChallengeMutationResultSerializer",
+    "ChallengeRatingSerializer",
+    "ChallengeSummarySerializer",
+    "ChallengeWriteSerializer",
+    "CtfScenarioListResponseSerializer",
+    "CtfScenarioRefSerializer",
+    "DeleteSuccessSerializer",
+    "EmailTemplateResponseSerializer",
+    "EmailTemplateRevertResultSerializer",
+    "EmailTemplateWriteSerializer",
+    "EventDetailSerializer",
+    "EventListResponseSerializer",
+    "EventMutationResultSerializer",
+    "EventSummarySerializer",
+    "EventWriteSerializer",
+    "FileDownloadResponseSerializer",
+    "FlagCreateResultSerializer",
+    "FlagWriteSerializer",
+    "ForceDeleteEventRequestSerializer",
+    "ForceDeleteEventResultSerializer",
+    "HintListResponseSerializer",
+    "HintWriteSerializer",
+    "NotificationAnnounceRequestSerializer",
+    "NotificationAnnounceResultSerializer",
+    "NotificationListItemSerializer",
+    "NotificationListResponseSerializer",
+    "NotificationSendResultSerializer",
+    "OrganizerChallengeDetailSerializer",
+    "OrganizerScoreboardResponseSerializer",
+    "ParticipantChallengeDetailSerializer",
+    "ParticipantChallengeFileSerializer",
+    "ParticipantChallengeListItemSerializer",
+    "ParticipantCurrentEventSerializer",
+    "ParticipantDeleteResultSerializer",
+    "ParticipantDetailSerializer",
+    "ParticipantEventSerializer",
+    "ParticipantHintSerializer",
+    "ParticipantImportErrorSerializer",
+    "ParticipantImportResultSerializer",
+    "ParticipantImportSerializer",
+    "ParticipantImportedItemSerializer",
+    "ParticipantInviteResultSerializer",
+    "ParticipantInviteSerializer",
+    "ParticipantListResponseSerializer",
+    "ParticipantRangeActionResultSerializer",
+    "ParticipantSelfSerializer",
+    "ParticipantSummarySerializer",
+    "ParticipantTeamMemberSerializer",
+    "ParticipantTeamSerializer",
+    "PrerequisiteCreateResultSerializer",
+    "PrerequisiteListResponseSerializer",
+    "PrerequisiteSerializer",
+    "PrerequisiteWriteSerializer",
+    "PublicScoreboardResponseSerializer",
+    "RangeAccessResponseSerializer",
+    "RangeListItemSerializer",
+    "RangeListResponseSerializer",
+    "RangeProvisionQueuedSerializer",
+    "RangeRecoveryRequestSerializer",
+    "RangeRecoveryResultSerializer",
+    "RangeStatusResponseSerializer",
+    "RateChallengeRequestSerializer",
+    "RateChallengeResultSerializer",
+    "ResendInviteResultSerializer",
+    "ScoreTimelineResponseSerializer",
+    "SendInvitationsResultSerializer",
+    "SparePoolRequestSerializer",
+    "SpareProvisionResultSerializer",
+    "SubmissionListItemSerializer",
+    "SubmissionListResponseSerializer",
+    "SubmitFlagRequestSerializer",
+    "SubmitFlagResultSerializer",
+    "UseHintRequestSerializer",
+    "UseHintResultSerializer",
+]
