+++ shifter/shifter_platform/ctf/api/serializers/_common.py
@@ -1,0 +1,12 @@
+"""Shared building-block serializers for the canonical CTF DRF API."""
+
+from __future__ import annotations
+
+from rest_framework import serializers
+
+
+class _NamedRefSerializer(serializers.Serializer):
+    """A minimal ``{id, name}`` reference to a related entity."""
+
+    id = serializers.CharField(read_only=True)
+    name = serializers.CharField(read_only=True)
