+++ shifter/shifter_platform/ctf/api/urls.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 
 from django.urls import path
 
-from ctf.api import organizer_views, participant_views, views
+from ctf.api import organizer, participant_views, views
 
 app_name = "ctf"
 
@@ -18,162 +18,162 @@ urlpatterns = [
         name="api_participant_challenge_detail",
     ),
     path("me/team/", participant_views.ParticipantTeamView.as_view(), name="api_participant_team"),
-    path("events/", organizer_views.EventListView.as_view(), name="api_event_list"),
-    path("events/<uuid:event_id>/", organizer_views.EventDetailView.as_view(), name="api_event_detail"),
+    path("events/", organizer.EventListView.as_view(), name="api_event_list"),
+    path("events/<uuid:event_id>/", organizer.EventDetailView.as_view(), name="api_event_detail"),
     path(
         "events/<uuid:event_id>/force-delete/",
-        organizer_views.ForceDeleteEventView.as_view(),
+        organizer.ForceDeleteEventView.as_view(),
         name="api_force_delete_event",
     ),
-    path("scenarios/", organizer_views.ScenarioListView.as_view(), name="api_scenarios"),
+    path("scenarios/", organizer.ScenarioListView.as_view(), name="api_scenarios"),
     path(
         "events/<uuid:event_id>/challenges/",
-        organizer_views.ChallengeListView.as_view(),
+        organizer.ChallengeListView.as_view(),
         name="api_challenge_list",
     ),
     path(
         "challenges/<uuid:challenge_id>/",
-        organizer_views.ChallengeDetailView.as_view(),
+        organizer.ChallengeDetailView.as_view(),
         name="api_challenge_detail",
     ),
-    path("challenges/<uuid:challenge_id>/submit/", organizer_views.SubmitFlagView.as_view(), name="api_submit_flag"),
-    path("challenges/<uuid:challenge_id>/hint/", organizer_views.UseHintView.as_view(), name="api_use_hint"),
+    path("challenges/<uuid:challenge_id>/submit/", organizer.SubmitFlagView.as_view(), name="api_submit_flag"),
+    path("challenges/<uuid:challenge_id>/hint/", organizer.UseHintView.as_view(), name="api_use_hint"),
     path(
         "challenges/<uuid:challenge_id>/hints/",
-        organizer_views.ChallengeHintsView.as_view(),
+        organizer.ChallengeHintsView.as_view(),
         name="api_challenge_hints",
     ),
-    path("hints/<uuid:hint_id>/delete/", organizer_views.HintDeleteView.as_view(), name="api_hint_delete"),
+    path("hints/<uuid:hint_id>/delete/", organizer.HintDeleteView.as_view(), name="api_hint_delete"),
     path(
         "challenges/<uuid:challenge_id>/rate/",
-        organizer_views.RateChallengeView.as_view(),
+        organizer.RateChallengeView.as_view(),
         name="api_rate_challenge",
     ),
-    path("submissions/", organizer_views.SubmissionListView.as_view(), name="api_submissions"),
+    path("submissions/", organizer.SubmissionListView.as_view(), name="api_submissions"),
     path(
         "events/<uuid:event_id>/participants/",
-        organizer_views.ParticipantListView.as_view(),
+        organizer.ParticipantListView.as_view(),
         name="api_participant_list",
     ),
     path(
         "events/<uuid:event_id>/participants/import/",
-        organizer_views.ParticipantImportView.as_view(),
+        organizer.ParticipantImportView.as_view(),
         name="api_participant_import",
     ),
     path(
         "participants/<uuid:participant_id>/",
-        organizer_views.ParticipantDetailView.as_view(),
+        organizer.ParticipantDetailView.as_view(),
         name="api_participant_detail",
     ),
     path(
         "participants/<uuid:participant_id>/resend-invite/",
-        organizer_views.ParticipantResendInviteView.as_view(),
+        organizer.ParticipantResendInviteView.as_view(),
         name="api_participant_resend_invite",
     ),
-    path("range/status/", organizer_views.ParticipantRangeStatusView.as_view(), name="api_range_status"),
-    path("range/access/", organizer_views.ParticipantRangeAccessView.as_view(), name="api_range_access"),
-    path("events/<uuid:event_id>/ranges/", organizer_views.EventRangeListView.as_view(), name="api_range_list"),
+    path("range/status/", organizer.ParticipantRangeStatusView.as_view(), name="api_range_status"),
+    path("range/access/", organizer.ParticipantRangeAccessView.as_view(), name="api_range_access"),
+    path("events/<uuid:event_id>/ranges/", organizer.EventRangeListView.as_view(), name="api_range_list"),
     path(
         "events/<uuid:event_id>/ranges/provision/",
-        organizer_views.EventRangeProvisionView.as_view(),
+        organizer.EventRangeProvisionView.as_view(),
         name="api_provision_ranges",
     ),
     path(
         "events/<uuid:event_id>/spares/",
-        organizer_views.EventSpareProvisionView.as_view(),
+        organizer.EventSpareProvisionView.as_view(),
         name="api_provision_event_spares",
     ),
     path(
         "participants/<uuid:participant_id>/range/provision/",
-        organizer_views.ParticipantRangeProvisionView.as_view(),
+        organizer.ParticipantRangeProvisionView.as_view(),
         name="api_provision_participant_range",
     ),
     path(
         "participants/<uuid:participant_id>/range/destroy/",
-        organizer_views.ParticipantRangeDestroyView.as_view(),
+        organizer.ParticipantRangeDestroyView.as_view(),
         name="api_destroy_participant_range",
     ),
     path(
         "participants/<uuid:participant_id>/range/stop/",
-        organizer_views.ParticipantRangeStopView.as_view(),
+        organizer.ParticipantRangeStopView.as_view(),
         name="api_stop_participant_range",
     ),
     path(
         "participants/<uuid:participant_id>/range/start/",
-        organizer_views.ParticipantRangeStartView.as_view(),
+        organizer.ParticipantRangeStartView.as_view(),
         name="api_start_participant_range",
     ),
     path(
         "participants/<uuid:participant_id>/range/restart/",
-        organizer_views.ParticipantRangeRestartView.as_view(),
+        organizer.ParticipantRangeRestartView.as_view(),
         name="api_restart_participant_range",
     ),
     path(
         "participants/<uuid:participant_id>/range/recover/",
-        organizer_views.ParticipantRangeRecoverView.as_view(),
+        organizer.ParticipantRangeRecoverView.as_view(),
         name="api_recover_participant_range",
     ),
     path(
         "participants/<uuid:participant_id>/bracket/",
-        organizer_views.AssignBracketView.as_view(),
+        organizer.AssignBracketView.as_view(),
         name="api_assign_bracket",
     ),
     path("events/<uuid:event_id>/scoreboard/", views.api_scoreboard, name="api_scoreboard"),
     path(
         "events/<uuid:event_id>/organizer-scoreboard/",
-        organizer_views.OrganizerScoreboardView.as_view(),
+        organizer.OrganizerScoreboardView.as_view(),
         name="api_organizer_scoreboard",
     ),
     path(
         "participants/<uuid:participant_id>/score-timeline/",
-        organizer_views.ScoreTimelineView.as_view(),
+        organizer.ScoreTimelineView.as_view(),
         name="api_score_timeline",
     ),
     path(
         "events/<uuid:event_id>/notifications/",
-        organizer_views.NotificationListView.as_view(),
+        organizer.NotificationListView.as_view(),
         name="api_notification_list",
     ),
     path(
         "notifications/<uuid:notification_id>/send/",
-        organizer_views.NotificationSendView.as_view(),
+        organizer.NotificationSendView.as_view(),
         name="api_notification_send",
     ),
     path(
         "events/<uuid:event_id>/email-templates/<str:notification_type>/",
-        organizer_views.EventEmailTemplateView.as_view(),
+        organizer.EventEmailTemplateView.as_view(),
         name="api_event_email_template",
     ),
     path(
         "events/<uuid:event_id>/invitations/send/",
-        organizer_views.SendInvitationsView.as_view(),
+        organizer.SendInvitationsView.as_view(),
         name="api_send_invitations",
     ),
     path(
         "challenges/<uuid:challenge_id>/flags/add/",
-        organizer_views.AddFlagView.as_view(),
+        organizer.AddFlagView.as_view(),
         name="api_add_flag",
     ),
-    path("flags/<uuid:flag_id>/remove/", organizer_views.RemoveFlagView.as_view(), name="api_remove_flag"),
+    path("flags/<uuid:flag_id>/remove/", organizer.RemoveFlagView.as_view(), name="api_remove_flag"),
     path(
         "challenges/<uuid:challenge_id>/files/",
-        organizer_views.ChallengeFilesView.as_view(),
+        organizer.ChallengeFilesView.as_view(),
         name="api_challenge_files",
     ),
     path(
         "files/<uuid:file_id>/delete/",
-        organizer_views.ChallengeFileDeleteView.as_view(),
+        organizer.ChallengeFileDeleteView.as_view(),
         name="api_challenge_file_delete",
     ),
-    path("files/<uuid:file_id>/download/", organizer_views.FileDownloadView.as_view(), name="api_file_download"),
+    path("files/<uuid:file_id>/download/", organizer.FileDownloadView.as_view(), name="api_file_download"),
     path(
         "challenges/<uuid:challenge_id>/prerequisites/",
-        organizer_views.ChallengePrerequisitesView.as_view(),
+        organizer.ChallengePrerequisitesView.as_view(),
         name="api_challenge_prerequisites",
     ),
     path(
         "prerequisites/<uuid:prerequisite_id>/delete/",
-        organizer_views.PrerequisiteDeleteView.as_view(),
+        organizer.PrerequisiteDeleteView.as_view(),
         name="api_prerequisite_delete",
     ),
 ]
